fixed some errors

+++ examples/009-iic-eeprom/f107vc/src/main.c
@@ -15,7 +15,7 @@ int main(void) {
     uint8_t data_to_read[1024];
     uint8_t data_to_compare[1024];
     for (uint16_t i = 0; i < 1024; i++) {
-        data_to_write[i] = ( uint8_t )(i + 10);
+        data_to_write[i] = (uint8_t)(i + 10);
     }
     uint16_t total_bytes_to_write = 200;
 
@@ -35,7 +35,7 @@ int main(void) {
     }
     uint32_t sum = 0;
     for (uint16_t i = 0; i < total_bytes_to_write; i++) {
-        data_to_compare[i] = data_to_read[i] - data_to_write[i];
+        data_to_compare[i] = (uint8_t)(data_to_read[i] - data_to_write[i]);
         sum += data_to_compare[i];
     }
     console.printf("compare sum = %d\r\n", sum);
@@ -48,13 +48,13 @@ int main(void) {
                    "%d bytes, no need to wait:\r\n",
                    __TIME__, total_bytes_to_write);
     for (uint16_t i = 0; i < 1024; i++) {
-        data_to_write[i] = ( uint8_t )(i + 15);
+        data_to_write[i] = (uint8_t)(i + 15);
     }
     eeprom.write.bytes(0, data_to_write, total_bytes_to_write);
     eeprom.read.bytes(0, data_to_read, total_bytes_to_write);
     sum = 0;
     for (uint16_t i = 0; i < total_bytes_to_write; i++) {
-        data_to_compare[i] = data_to_read[i] - data_to_write[i];
+        data_to_compare[i] = (uint8_t)(data_to_read[i] - data_to_write[i]);
         sum += data_to_compare[i];
     }
     console.printf("compare sum = %d\r\n", sum);
