+++ one-third-hal/core/uart/usart2.c
@@ -139,20 +139,20 @@ static void Usart2IdleIrqCallback(void) {
         static uint16_t remaining[2] = { 0, 0 };
         uint16_t receive_count = 0;
         remaining[0] = remaining[1];
-        remaining[1] = ( uint16_t )(__HAL_DMA_GET_COUNTER(usart2.huart.hdmarx));
+        remaining[1] = (uint16_t)(__HAL_DMA_GET_COUNTER(usart2.huart.hdmarx));
         static bool dma_first_packet = true;
         if (dma_first_packet) {
-            receive_count = usart2.rb.state.capacity - remaining[1];
+            receive_count = (uint16_t)(usart2.rb.state.capacity - remaining[1]);
             dma_first_packet = false;
             op.ringbuffer.pushN(&(usart2.rb), usart2.rb.data, receive_count);
         }
         else {
             if (remaining[0] > remaining[1]) {
-                receive_count = remaining[0] - remaining[1];
+                receive_count = (uint16_t)(remaining[0] - remaining[1]);
             }
             else {
-                receive_count = ( uint16_t )(usart2.rb.state.capacity
-                                             + remaining[0] - remaining[1]);
+                receive_count = (uint16_t)(usart2.rb.state.capacity
+                                           + remaining[0] - remaining[1]);
             }
             // DMA has already pushed data into the ringbuffer
             op.ringbuffer.added(&(usart2.rb), receive_count);
